Feature - project collab perms
* bumped schema with addition of collab role
* moved cert router to start of person
* added sendResponse method to terminate execution early
