+++ lib/api.js
@@ -33,6 +33,7 @@ module.exports = settings => {
 
   app.use(require('./middleware/user'));
   app.use(require('./middleware/permissions-bypass'));
+  app.use(require('./middleware/send-response'));
 
   app.use('/search', proxy(`${settings.search}`));
 
@@ -42,33 +43,7 @@ module.exports = settings => {
   app.use('/establishment(s)?', establishmentRouter(settings));
   app.use('/task(s)?', require('./routers/task'));
 
-  app.use((req, res, next) => {
-    if (!req.permissionChecked) {
-      const nopes = ['/tasks', '/me'];
-      if (!nopes.includes(req.path)) {
-        req.log('info', { url: req.originalUrl, event: 'unchecked-permissions' });
-      }
-    }
-    if (res.response !== undefined) {
-      const response = {};
-      if (!req.query.onlymeta) {
-        response.data = res.response;
-      }
-      response.meta = Object.assign({}, res.meta);
-      if (req.establishment) {
-        response.meta.establishment = {
-          id: req.establishment.id,
-          name: req.establishment.name,
-          status: req.establishment.status,
-          revocationDate: req.establishment.revocationDate,
-          issueDate: req.establishment.issueDate,
-          isTrainingEstablishment: req.establishment.isTrainingEstablishment
-        };
-      }
-      return res.json(response);
-    }
-    next();
-  });
+  app.use((req, res) => res.sendResponse());
 
   app.use((req, res, next) => {
     next(new NotFoundError());
