+++ lib/routers/profile/certificates.js
@@ -73,13 +73,13 @@ app.put('/:certificateId', permissions('training.update', req => ({ profileId: r
 app.delete('/:certificateId', permissions('training.update', req => ({ profileId: req.profileId })), submit('delete'));
 app.post('/', validateCertificate(), permissions('training.update', req => ({ profileId: req.profileId })), submit('create'));
 
-app.get('/', permissions('training.read', req => ({ profileId: req.profileId })), (req, res, next) => {
+app.get('/', permissions('training.read', req => ({ ...(req.query || {}), profileId: req.profileId })), (req, res, next) => {
   const { Certificate } = req.models;
   Certificate.query().where({ profileId: req.profileId })
     .then(certificates => {
       res.response = certificates;
+      res.sendResponse();
     })
-    .then(() => next())
     .catch(next);
 });
