+++ lib/routers/profile/person.js
@@ -150,8 +150,10 @@ const getSingleProfile = req => {
         .then(() => req.user.can('profile.read.basic', req.params))
         .then(allowed => {
           if (allowed) {
+            console.log({ path: req.path });
             return profileQueries.getNamed();
           }
+
           throw new NotFoundError();
         });
     })
@@ -249,6 +251,8 @@ const getPil = (req, res, next) => {
 module.exports = (settings) => {
   const router = Router({ mergeParams: true });
 
+  router.use('/certificate(s)?', require('./certificates'));
+
   router.use((req, res, next) => {
     Promise.resolve()
       .then(() => getSingleProfile(req))
@@ -299,8 +303,6 @@ module.exports = (settings) => {
     },
     fetchOpenTasks()
   );
-
-  router.use('/certificate(s)?', require('./certificates'));
 
   return router;
 };
